Code refaction of get_home_diff function
Improve the function description, now change the global variable instead
of returning a string with echo. Also the diff line is decoupled into
the diff between home_before and after, and the sed part to clean the
paths.

Now the app_dirs_and_files variable is converted to an array as expected
by other functions in the script.

Also some there are some minor improvements to the output of the
function (like when printing the elements of app_dirs_and_files).